refactor(scanner): extract draw.io XML primitives to diagram_drawio.py
diagram-gen.py (930 lines) mixed low-level mxGraph/draw.io element builders with
the high-level page/diagram composition. The seven primitives — mx_escape,
drawio_cell, emit_mx_geometry, create_drawio_root, create_multipage_drawio_root,
add_drawio_page, _draw_edge — are a pure LEAF cluster: they use only
xml.etree.ElementTree and each other, with zero back-dependency on the scanner's
aggregate data / CATEGORIES / ARCH_DOMAINS / the generate_* functions.

Hoist them to a new leaf module scanner/lib/diagram_drawio.py and import them
back into diagram-gen.py. No circular import (diagram_drawio does not import
diagram-gen); the imported names remain diagram-gen module attributes, so the
page builders below and the tests (MOD.mx_escape / MOD.drawio_cell …) resolve
them exactly as before. diagram-gen.py drops ~90 lines (930 -> ~840).

Priority-2 step 2 of docs/plans/refactoring-backlog.md.

Verified — behavior-preserving, coverage-neutral: pytest 1443 passed,
scanner/lib coverage 99.12% (>=99% gate); the diagram suites (canonical-sync
guard + pure helpers + smoke) 112 passed; ast.parse OK on both files. The
canonical-sync guard (ARCH_DOMAINS / COMPLIANCE_FRAMEWORKS / CATEGORIES
single-sourcing) still holds — those constants were untouched.
